api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-pymongo>=3.0.1",
    "pydantic>=2.11.3",
]
